hand_written_recog/requirements.txt
#python==3.10
Flask==2.2.5
torch==2.5.1
torchvision==0.20.1
Pillow
albumentations
opencv-python
matplotlib
